+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   * Verify peer certificate for SSL (only available in OTP 17.3)
   * Reduce archive size with compiler option `debug_info: false`
   * Add support for config as an erlang term file
+  * Warn if Hex was built against a different major.minor Elixir version
 
 ## v0.4.3 (2014-09-06)